Move Cloud.py into subdir to fix Sphinx issue
