+++ tinytuya/tuyacloud/__init__.py
@@ -1,0 +1,2 @@
+
+from .Cloud import Cloud
